Carl/bcda 8967 adjust deploy verifications (#1117)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8967

## 🛠 Changes

Convert deploy workflow version verifications to run in parallel with a
long retry interval to account for lengthy instance refreshes.

## ℹ️ Context

Currently we are just sleeping for 240s then checking versions. This is
failing deploys a bit more than 50% of the time.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Test deploy run on PR branch push.