BCDA-9067: Adjust loggers to include transaction_id for SSAS (#1119)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9067

## 🛠 Changes

Added transaction id across a few different loggers as well as to auth
requests to SSAS.

## ℹ️ Context

Currently our logging of auth requests is lacking, often times we are
seeing nothing when auth requests fail.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.