BCDA-8967: Fix workflow deploy scripts (#1120)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8967

## 🛠 Changes

Slight adjustments to deploy workflow, messed up the env var name, and
need to remove the one exit clause.

## ℹ️ Context

Fix deploy workflow verification checks. This work is a slight
adjustment to previous work.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Workflow testing