Merge branch 'carl/BCDA-8967-add-env-to-jobs' (#1125)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8967

## 🛠 Changes

Update jq query to catch for malformed json.

## ℹ️ Context

Currently if the page is not available it will not return json, which
will crash the script.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->