BCDA-9102: Make sure config is available for worker instance (#1130)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9102

## 🛠 Changes

Not represented in this PR:
- Added param store values for BCDA_WORKER_CONFIG_PATH to point at
(cloned) worker config.yml files (also in param store). This doubles the
config.yml files we have to keep in sync but was the easiest path
forward. (The alternative is messing around with an archaic script in
bcda-ops, something that we have spent a lot of time trying to refactor
as part of our AL2023 work).

In this PR:
- Add error checks for when we fail to load ACO Config appropriately.
- Add tests for the above and update existing.
- Some cleanup of now unused functions.

## ℹ️ Context

Previously if we failed to load ACO Configs we would just silently
ignore the related logic blocks. This included important logic checks.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.