BCDA-9018: demo endpoints (v3) initial stubbing out (#1132)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9018

## 🛠 Changes

Duplicating v2 routes at /demo namespace to demo out v3 requests to BFD.
Gated by feature flag. Removal of Claim and Claim Response for v3
requests. Initial stubbing out of typefilter (commented out for now).

## ℹ️ Context

For the Connectathon we want a barebones functionality for v3 requests
to BFD.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.

---------

Co-authored-by: Colby Seyferth <colbyseyferth@navapbc.com>