BCDA-9002: migrate workflows to GF (#1118)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9002
https://jira.cms.gov/browse/BCDA-9091

## 🛠 Changes

- account ID pointing at GF non-prod (prod to be added after testing)
- combining workflow files for lambdas that deploy to multiple envs;
conditional added for env selection and auto deploy env

## ℹ️ Context

Part of GF migration. For the remaining workflows that need to be
migrated, please see BCDA-9143.

Three workflows are currently unable to be tested due to resolving
infra/network related issues in the dev env.

1. admin-create-aco-creds-integration-test-GF.yml (this one may be a
permission issue if problem persists)
2. fhirscan-GF.yml
3. smoke-test-GF.yml

Not able to test inputs unless the file has been merged or already
exists, so once these changes are merged we can validate the inputs work
as expected.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tested workflows by running them. All except for the aforementioned
three have been verified and one known/expected failure: Create-Group
does not do cleanup (BCDA-9148) so it will fail on all runs after the
first.