PLT-891 BCDA GHA build-ami for GF (#1100)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-963

## 🛠 Changes

### Code changes:
- Updated GHA runs-on to codebuild.
- Created matrix to run workflow on both PROD and NON-PROD accounts
simultaneously using the CDAP VPC in PROD and the BCDA DEV VPC in
NON-PROD.
- Updated GHA role-to-assume to
'role/delegatedadmin/developer/bcda-[prod|dev]-github-actions' - prod
for PROD and dev for NON-PROD.
- Updated Ansible install step to use DNF and v8.3.0 (as per Julia's
legacy PR) so that ansible-playbook is on path.
- Reverted tmp dir from /home/ec2-user/ to /tmp/ - this could change
depending on gold image ultimately used.
- Updated Packer to connect to EC2 instance via SSM instead of SSH due
to AL2023 changes.

### Non-code changes:
- Created Github bcda-app repo secret via GH cli for PROD_ACCOUNT_ID and
NON_PROD_ACCOUNT_ID.
- Updated GF PROD account parameter '/bcda/workflows/packer_subnet_id'
to ID of cdap-east-mgmt private subnet in PROD and bcda-east-dev private
subnet in NON-PROD.
- Updated GF PROD account parameter '/bcda/workflows/gold_image_ami' to
AMI ID corresponding to name al2023-gi-2025-04-09T10-40-28Z.
- Manually recreated 'bcda-packer' AWS IAM role and instance profile on
GF PROD and NON-PROD accounts.
- Manually recreated 'packer_sg' security group on GF PROD and NON-PROD
accounts, but now only with one inbound rule allows SSH traffic from
either the CDAP VPC in PROD and the BCDA DEV VPC in NON-PROD (depending
on what account the SG was created in). Even though Packer now connects
to the builder instance via SSM it still requires this security group
and rule or else it will fail with a timeout waiting for SSH to become
ready.

## ℹ️ Context

These changes are required for the move of the BCDA API and worker
instances to GF.

## 🧪 Validation

The following GHA run successfully created platinum AMIs in the GF PROD
and NON-PROD accounts:
https://github.com/CMSgov/bcda-app/actions/runs/14780942497

---------

Co-authored-by: Carl Partridge <carlpartridge@navapbc.com>