BCDA-9069: update fhir_testing to use alpine image and non-root user (#1135)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9069

## 🛠 Changes

Updated the dockerfile to:
- use the ubuntu image instead of the ruby one. There's no need for ruby
(initially was thinking about running the inferno server on it which is
ruby, but have gone away from that plan)
- Create a non-root user to use for running the inferno tests, so we
don't use the root user

## ℹ️ Context

it's better practice to not use a root user, for security reasons.
Because of this, SonarQube was complaining too. So instead of using a
root user, we'll create a new user with limited permissions and use that
one for running the inferno scripts.

## 🧪 Validation

Regression: Run the [FHIR Scan
workflow](https://github.com/CMSgov/bcda-app/actions/workflows/fhirscan.yml)
and make sure that still works

See that the sonarqube checks passed on this PR/branch