BCDA-9148: Migrate and add cleanup step to create group lambda (#1145)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9143
https://jira.cms.gov/browse/BCDA-9148

## 🛠 Changes

Update create group lambda deployment workflow to use codebuild runners
Add a cleanup step to the create group lambda integration test that
deletes the group row that is created during the test.

## ℹ️ Context

This workflow does not do any database cleanup after the test runs, so
it will fail on the second run. This PR should add some logic in the
workflow to delete the row in the database when/if it's been created. It
also checks that the number of rows to be deleted is equal to one for a
safer operation.

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

Check that the test workflow passes, especially after a rerun.