BCDA-9287: remove init from response writer (#1170)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9287

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This PR refactors the responseutils writer.go to move the logic present
in its `init()` function to a constructor.
The json marshaller that was instantiated just once in the init()
statement is now part of the state of the ResponseWriter type. As a
result, other functions in this file were refactored to be receiver
functions of the ResponseWriter, and whenever you need to call any of
these functions, a ResponseWriter must be created with a json
marshaller.


## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->