BCDA-9286-remove-bcda-queue (#1172)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9286

## 🛠 Changes

- Remove bcda_queue from
  - dbdocs
  - Makefile
  - docker-compose (and test)
  - migration readme
  - integration tests
  - lambda tests
  - config (and test)

## ℹ️ Context

We do not use the bcda_queue database anymore. This was utilized by the
que-go package. Once we moved to river for job processing we created
tables under the bcda DB and consolidated our databases. Que-go package
has been removed earlier and this PR is addressing the removal of all
bcda_queue related items in the repo.

## 🧪 Validation

Locally build, lint, and run tests on bcda-app