BCDA-9286-fix-drop-database-call (#1173)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9286

## 🛠 Changes

~Refactored the DB migration and took out DROP DATABASE from the
function. We are still closing connections and dropping only if they
exist~

Removed the migrations completely, and we can remove the DB manually.
This is due to a limitation with the golang-migrate tool running all
migrations in transactions by default and Postgres not allowing
CREATE/DROP within transactions.

## ℹ️ Context

bcda_queue DB is obsolete, and we have consolidated all tables under
bcda DB

## 🧪 Validation

Local lint, build, and test