BCDA-9314: refactor API db connection globals (#1171)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9314
https://jira.cms.gov/browse/BCDA-9287

## 🛠 Changes

- moved database connection init logic to `connect` functions, which
establish new connections to the database and replace init and global
connection variables
- split db connection and pool into separate `connect` functions with
separate health checks
- refactored files that relied on global vars (e.g. routers, middleware,
apis) to structs with explicitly defined dependencies on database
connections in apis, routers, and services
- configured database connections at code entrypoints (e.g. cli setup)
and passed connections down to services
- refactored `auth/Provider.go` to not be a global, either, and instead
be configured at setup and passed down like the database connections

To-Do:

- [X] refactor references to global connection in lambdas
- [X] refactor references to global connection in various other places
(middleware, health check, etc.)
- [X] remove global connection var
- [x] add a unit test for pool connection check

## ℹ️ Context

This PR demonstrates a strategy to refactor the database Connection --
specifically, to remove the global Connection variables defined in
/database/connection.go which are created in an init() function, and
instead do the following:
1. Create the database connection(s) once at each code entrypoint (e.g.
cli.go, each lambda main.go, etc.)
2. Pass each connection as a dependency to the routers and services that
need it.
3. For each router/service that needs a connection, create a type with
state that stores the connection, as well as a constructor that accepts
its required dependencies, if necessary.

As a result, many other files needed to be changed to store the db
connection within its state, including the API files and middleware.
Notably, the Provider, which also depends on the db connection, was also
being used as a global across many files, so it too had to be updated to
a stateful type which is constructed at entrypoints.
Finally, many tests were updated to account for these changes. The
majority of the test updates were completely innocuous -- replacing
references to a global connection with references to a constructed
connection -- but others involved untangling hidden dependencies on the
database, provider, etc. that became very obvious after they were
refactored.

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->