BCDA-9298: updates to logging for data endpoint (#1183)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9298

## 🛠 Changes

added additional context and transaction ID to logs in data endpoint.

## ℹ️ Context

Added for more visibility into data endpoint and to be able to correlate
logs.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->