BCDA-9319: Fix workflows for all envs for aurora (#1184)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9319

## 🛠 Changes

Adjusted two workflows to allow for each env, they were previously
hardcoded to dev.

## ℹ️ Context

Recent changes to our DB (migration to Aurora) required changes in our
workflows that talked directly to our DB. Most of these are integration
tests that only ever talk to one env (eg `test`).

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Successful run of one of the workflows for test env:
https://github.com/CMSgov/bcda-app/actions/runs/16880547634/job/47815071171