BCDA-9355: Remove init from servicemux package (#1189)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9355

## 🛠 Changes

remove init() from `servicemux` package.

## ℹ️ Context

Generous init() usage across the repo often makes testing cumbersome.
This change is part of the effort to remove init() across the project to
simplify testing.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests pass and were modified.