BCDA-9374: Remove resourcetype init (#1195)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9374

## 🛠 Changes

- Removed init and put variable initialization as unexported package
level var
- Renamed `DataType` to `ClaimType`
- Renamed `dataTypeMap` to `fhirResourceTypeMap`

## ℹ️ Context

Main changes that aren't part of variable rename are in `resource.go`.
Part of the ongoing effort to remove init() across the project to make
development and testing faster + more reliable.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests pass.