BCDA-9367: Add new auto deploy to dev workflow (#1196)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9367

## 🛠 Changes

Added a new workflow to address issues with the daily dev deploy from
main.

## ℹ️ Context

The daily dev deploy to main has been having troubles since our
greenfield migration. We are no longer getting the latest changes
deployed to dev daily. Adding this new workflow should address this. It
felt better to create a new workflow then to try to finagle existing
workflows (which would have involved a lot of if else checks and such).

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Testing workflow in PR pushes. (Successful run:
https://github.com/CMSgov/bcda-app/actions/runs/17082274241)