BCDA-9415: Update v3 capabilitystatement (#1210)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9415

## 🛠 Changes

Updated the V3 CapabilityStatement:
- [x] Add Parameter documentation for _type and _since for both Patient
and Group resources
- [x] update the operation name for patient-export and group-export
(change to just "export")
- [x] Add _typeFilter parameter details for ExplanationOfBenefit _tag
(adjudication-level) filtering

## ℹ️ Context

Guidance changes in the STU3 veresion of the Bulk Data IG to correct for
the operation name. This updates our capabilityStatement to have the
correct value for v3.

In addition to updating the operation name, we were also lacking
documentation about our supported filtering parameters. Added
documentation for _type, _since, _typeFilter, and the _typeFilter
subquery: ExplanationOfBenefit._tag parameter

## 🧪 Validation

Hit the `/metadata` and confirm the new elements are returned in the
CapabilityStatement resource for only the demo/v3 endpoint

v1 and v2 should be unchanged.

unit tests pass

---------

Co-authored-by: Parwinder Bhagat <Parwinder.Bhagat@e14s.com>