BCDA-9392: Test Container Database (#1207)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9392

## 🛠 Changes

**TestDatabaseContainer type that wraps golang
[testcontainers](https://golang.testcontainers.org/modules/postgres/)
package**. Why?
- BCDA api is using pgx, pgx pool, and sql/db for database connections;
wrapper will create a connection for any of these, which eliminates the
need to create a connection in each test suite/file.
- Calling NewTestDatabaseContainer ensures that the same exact container
setup is called for each test, each test suite, or each file and removes
the burden of configuring for the developer.
- The wrapper has methods such as ExecuteFile and ExecuteDir, which will
help enforce the location and naming of testdata across the repository.
- README has been created for example and intended usage.

**api_test.go used for initial tests to apply containers**
- Example file/suite that was used to try the test container.
- TestJobStatusNotComplete was the primary test that was updated/tested
for the containers

**db/testdata directory creation**
- Standard directory naming for test data.
- README created for example and intended usage

## ℹ️ Context

**The goal of this PR is to**:
- set a standard that can be used across all tests, which will improve
readability and speed up development time
- implement a solution that can be incrementally rolled out with other
dev work with no disruption to current tests or CI/CD.
- reduce development burden when creating tests
- improve test reliability, since the database will not have potential
state changes in between tests

Currently, a single container database is used across all tests, with
some or no cleanup, depending on the tests. There is a risk that tests
are not running in a "clean" state; the shared database could be altered
from a previous test which can affect the accuracy and reliability of
all tests.

Additionally, there is no standard of how we do database integration
tests or unit tests. For database integration tests, we should have the
same setup for the database every time, across all tests that will
utilize the database. This will ensure that the tests are in a good
state before running and their outcome will not affect other tests.

For unit tests, we should be mocking the database each time, but this PR
only address integration test strategies. Because this container
strategy would be implemented at whatever granularity is desired (test
file, suite, sub test, etc) and does not change the current test
database that is used across all tests, it will not affect our current
test suite and our tests can be gradually updated over time, as
developers are working in various package.

With the new changes, a test database container will be created and each
time a new test runs, the database will be restored back to a good known
state with a snapshot, which will ensure a clean state for each run.
Testing data and additional snapshots can be created as needed,
depending on the testing scenario.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests passin, v2/test_api.go file updated to use new test containers
(arbitrarily chosen), and tests added for container.go.

---------

Co-authored-by: Parwinder Bhagat <Parwinder.Bhagat@e14s.com>