BCDA-9088: update config files for IOTA model (#1227)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9088

## 🛠 Changes

- Updated dev & test config files and sops files to include IOTA model
Note: reverted changes to read from the config file - not picking up the
env file during testing. Creating a follow up ticket so we don't have to
hardcode new models into `service.IsSupportedACO`.


## ℹ️ Context

New models need to be added to the config files to grant access to the
application to entities under that model.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests passing, deployed dev branch for create-aco lambda and verified I
can add an IOTA test entity.

---------

Co-authored-by: Parwinder Bhagat <Parwinder.Bhagat@e14s.com>