BCDA-9362: speed up loading test data (#1229)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9362

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
Run a docker container once and exec into it to run scripts instead of
booting multiple containers one after another

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

The previous PR in the repo for this ticket slowed down an already-slow
`make docker-bootstrap`
This PR speeds it up to be much faster than even before the previous PR
by booting the API container once and exec-ing into it to perform data
seeding instead of booting api containers one after another.
This may also prevent all or most of the many orphan containers that
`make docker-bootstrap` previously created.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->