BCDA-9263: Adjust deploy workflows a bit (#1230)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9263

## 🛠 Changes

Small adjustments to workflows:
- remove multiple calls to ci_checks workflow as it was both running the
same lengthy CI checks multiple times as well as causing an error on the
code-coverage results artifact.

## ℹ️ Context

Changes to various workflows are often times hard to test outside of
running them in dev environment.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Re-running workflows in dev.